Move config creation logic into config module.
