+++ src/config.ts
@@ -9,8 +9,16 @@
  * rights grant found at http://polymer.github.io/PATENTS.txt
  */
 
+import * as fsExtra from 'fs-extra';
 import * as path from 'path';
-import {CheckConfig} from './github';
+
+import {fcpBrowsers} from './browser';
+import {Config} from './config';
+import {parseConfigFile, writeBackSchemaIfNeeded} from './configfile';
+import * as defaults from './defaults';
+import {Opts} from './flags';
+import {CheckConfig, parseGithubCheckFlag} from './github';
+import {specsFromOpts} from './specs';
 import {Horizons} from './stats';
 import {BenchmarkSpec} from './types';
 import {fileKind} from './versions';
@@ -31,6 +39,97 @@ export interface Config {
   remoteAccessibleHost: string;
   forceCleanNpmInstall: boolean;
   csvFile: string;
+}
+
+export async function makeConfig(opts: Opts): Promise<Config> {
+  // These options are only controlled by flags.
+  const baseConfig = {
+    mode: (opts.manual === true ? 'manual' : 'automatic') as
+        ('manual' | 'automatic'),
+    savePath: opts.save,
+    githubCheck: opts['github-check'] ?
+        parseGithubCheckFlag(opts['github-check']) :
+        undefined,
+    remoteAccessibleHost: opts['remote-accessible-host'],
+  };
+
+  let config: Config;
+  if (opts.config) {
+    if (opts.root !== undefined) {
+      throw new Error('--root cannot be specified when using --config');
+    }
+    if (opts.browser !== undefined) {
+      throw new Error('--browser cannot be specified when using --config');
+    }
+    if (opts['sample-size'] !== undefined) {
+      throw new Error('--sample-size cannot be specified when using --config');
+    }
+    if (opts.timeout !== undefined) {
+      throw new Error('--timeout cannot be specified when using --config');
+    }
+    if (opts.horizon !== undefined) {
+      throw new Error('--horizon cannot be specified when using --config');
+    }
+    if (opts.measure !== undefined) {
+      throw new Error('--measure cannot be specified when using --config');
+    }
+    if (opts['resolve-bare-modules'] !== undefined) {
+      throw new Error(
+          '--resolve-bare-modules cannot be specified when using --config');
+    }
+    if (opts['window-size'] !== undefined) {
+      throw new Error('--window-size cannot be specified when using --config');
+    }
+    const rawConfigObj = await fsExtra.readJson(opts.config);
+    const validatedConfigObj = await parseConfigFile(rawConfigObj);
+
+    await writeBackSchemaIfNeeded(rawConfigObj, opts.config);
+
+    config = {
+      ...baseConfig,
+      ...validatedConfigObj,
+    };
+
+  } else {
+    config = {
+      ...baseConfig,
+      root: opts.root !== undefined ? opts.root : defaults.root,
+      sampleSize: opts['sample-size'] !== undefined ? opts['sample-size'] :
+                                                      defaults.sampleSize,
+      timeout: opts.timeout !== undefined ? opts.timeout : defaults.timeout,
+      horizons: parseHorizons(
+          opts.horizon !== undefined ? opts.horizon.split(',') :
+                                       defaults.horizons),
+      benchmarks: await specsFromOpts(opts),
+      resolveBareModules: opts['resolve-bare-modules'] !== undefined ?
+          opts['resolve-bare-modules'] :
+          true,
+      forceCleanNpmInstall: opts['force-clean-npm-install'],
+      csvFile: opts['csv-file'],
+    };
+  }
+
+  if (config.sampleSize <= 1) {
+    throw new Error('--sample-size must be > 1');
+  }
+
+  if (config.timeout < 0) {
+    throw new Error('--timeout must be >= 0');
+  }
+
+  if (config.benchmarks.length === 0) {
+    throw new Error('No benchmarks matched with the given flags');
+  }
+
+  for (const spec of config.benchmarks) {
+    if (spec.measurement === 'fcp' && !fcpBrowsers.has(spec.browser.name)) {
+      throw new Error(
+          `Browser ${spec.browser.name} does not support the ` +
+          `first contentful paint (FCP) measurement`);
+    }
+  }
+
+  return config;
 }
 
 /**
